+++ src/main/java/com/nisovin/shopkeepers/config/lib/Config.java
@@ -22,8 +22,8 @@ import com.nisovin.shopkeepers.config.lib.annotation.WithValueTypeProvider;
 import com.nisovin.shopkeepers.config.lib.setting.FieldSetting;
 import com.nisovin.shopkeepers.config.lib.setting.Setting;
 import com.nisovin.shopkeepers.config.lib.value.DefaultValueTypes;
-import com.nisovin.shopkeepers.config.lib.value.UnknownEntityTypeException;
-import com.nisovin.shopkeepers.config.lib.value.UnknownMaterialException;
+import com.nisovin.shopkeepers.config.lib.value.InvalidEntityTypeException;
+import com.nisovin.shopkeepers.config.lib.value.InvalidMaterialException;
 import com.nisovin.shopkeepers.config.lib.value.ValueLoadException;
 import com.nisovin.shopkeepers.config.lib.value.ValueType;
 import com.nisovin.shopkeepers.config.lib.value.ValueTypeProvider;
@@ -360,9 +360,9 @@ public abstract class Config {
 	protected <T> void onValueLoadException(ConfigurationSection config, Setting<T> setting, ValueLoadException e) throws ConfigLoadException {
 		String configKey = setting.getConfigKey();
 		Log.warning(this.msgValueLoadException(configKey, e));
-		if (e instanceof UnknownMaterialException) {
+		if (e instanceof InvalidMaterialException) {
 			Log.warning(this.getLogPrefix() + "All valid material names can be found here: https://hub.spigotmc.org/javadocs/bukkit/org/bukkit/Material.html");
-		} else if (e instanceof UnknownEntityTypeException) {
+		} else if (e instanceof InvalidEntityTypeException) {
 			Log.warning(this.getLogPrefix() + "All valid entity type names can be found here: https://hub.spigotmc.org/javadocs/bukkit/org/bukkit/entity/EntityType.html");
 		}
 	}