+++ src/main/java/com/nisovin/shopkeepers/config/lib/value/InvalidEntityTypeException.java
@@ -5,15 +5,15 @@ import org.bukkit.entity.EntityType;
 /**
  * This exception is thrown by {@link ValueType#load(Object)} when an unknown {@link EntityType} is encountered.
  */
-public class UnknownEntityTypeException extends ValueLoadException {
+public class InvalidEntityTypeException extends ValueLoadException {
 
 	private static final long serialVersionUID = 4248498809095698671L;
 
-	public UnknownEntityTypeException(String message) {
+	public InvalidEntityTypeException(String message) {
 		super(message);
 	}
 
-	public UnknownEntityTypeException(String message, Throwable cause) {
+	public InvalidEntityTypeException(String message, Throwable cause) {
 		super(message, cause);
 	}
 }