+++ src/main/java/com/nisovin/shopkeepers/config/lib/value/types/MaterialValue.java
@@ -2,7 +2,7 @@ package com.nisovin.shopkeepers.config.lib.value.types;
 
 import org.bukkit.Material;
 
-import com.nisovin.shopkeepers.config.lib.value.UnknownMaterialException;
+import com.nisovin.shopkeepers.config.lib.value.InvalidMaterialException;
 import com.nisovin.shopkeepers.config.lib.value.ValueLoadException;
 import com.nisovin.shopkeepers.config.lib.value.ValueParseException;
 
@@ -15,7 +15,18 @@ public class MaterialValue extends MinecraftEnumValue<Material> {
 	}
 
 	@Override
-	protected ValueLoadException newUnknownEnumValueException(String valueName, ValueParseException parseException) {
-		return new UnknownMaterialException(parseException.getMessage(), parseException);
+	protected ValueLoadException newInvalidEnumValueException(String valueName, ValueParseException parseException) {
+		return new InvalidMaterialException(parseException.getMessage(), parseException);
+	}
+
+	@Override
+	public Material parse(String input) throws ValueParseException {
+		Material material = super.parse(input);
+		assert material != null; // Parsing throws an exception instead
+		// Filter legacy materials:
+		if (material.isLegacy()) {
+			throw new ValueParseException("Unsupported legacy Material: " + input);
+		}
+		return material;
 	}
 }