+++ CHANGELOG.md
@@ -138,6 +138,7 @@ However, if the shopkeeper actually moved from its previous spawn location and n
 * Performance: Improvements related to checking and deleting shopkeepers of inactive players.
 * Metrics: The shopkeepers count chart groups its results into slightly more detailed categories now.
 * Added editor option to toggle the invulnerability of regular villagers.
+* Config: Legacy item types, and item types that are not actually items, can no longer be specified. We log a warning now when we encounter such an item type, similar to how we deal with unknown item types.
 
 API:  
 * Renamed TradingOffer to TradeOffer.